prep for pull to get config/run update
