+++ .gitignore
@@ -12,4 +12,4 @@ node_modules
 .DS_Store
 temp
 
-.idea/
+/.idea
