+++ api/database.js
@@ -1,9 +1,9 @@
+var logger = require('./logging');
+var config = require('./config');
+
 var Bookshelf = require('bookshelf');
 var Knex = require('knex');
 var milliseconds = require('ms');
-
-var logger = require('./logging');
-var config = require('./config');
 
 var KNEX_CONFIG = {
 	client: 'mysql',
@@ -23,6 +23,8 @@ var KNEX_CONFIG = {
 
 function DatabaseManager() {
 	this._knex = Knex(KNEX_CONFIG);
+	this._knex = (config.isTest) ? require('mock-knex').mock(this._knex) : this._knex;
+
 	this._bookshelf = Bookshelf(this._knex);
 }
 
@@ -35,6 +37,9 @@ DatabaseManager.prototype.connection = function () {
 	return this._knex;
 };
 
-logger.info("connected to database as %s", config.database.primary.user);
+if (!config.isTest) {
+	// do not say this when the datastore is mocked
+	logger.info("connected to database as %s", config.database.primary.user);
+}
 
 module.exports = new DatabaseManager();
