+++ test/auth/hacker.js
