+++ test/user/admin.js
