+++ test/user/hacker.js
